+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `tanjun.dependencies.inject_lc(Type)` which is a shorthand for
+  `tanjun.injected(callback=tanjun.make_lc_resolver(Type))`.
+- `tanjun.dependencies.cached_inject(...)` which is a shorthand for
+  `tanjun.injected(callback=tanjun.cache_callback(...))`.
+- `tanjun.inject` which is identical to `tanjun.injecting.injected` but does not replace it.
+
+### Fixed
+- Removed `cache_callback` from injecting.pyi
 
 ## [2.1.2a1] - 2021-10-15
 ### Added